backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0
prophet==1.1.4
tensorflow==2.14.0
python-dotenv==1.0.0
sqlalchemy==2.0.22
pydantic==2.4.2
pytest==7.4.2
httpx==0.25.0
joblib==1.3.2
plotly==5.17.0
pymongo==4.5.0
